refactor: leader-micrometer 의존성 leader-spring-boot-common → leader-core 교체 (Phase B T8-T11)
